requirements.txt
absl-py==2.1.0
attrs==23.2.0
cffi==1.16.0
contourpy==1.1.1
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.49.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
jax==0.4.13
kiwisolver==1.4.5
matplotlib==3.7.5
mediapipe-nightly==0.10.10.post20240212
ml-dtypes==0.2.0
MouseInfo==0.1.3
numpy==1.24.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
protobuf==3.20.3
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.1.0
scipy==1.10.1
six==1.16.0
sounddevice==0.4.6
zipp==3.17.0
